requirements.txt
streamlit==1.31.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
plotly==5.18.0
PyPDF2==3.0.1
pymongo==4.6.1